Make compatible with cropduster 4 fields
Apparently reversion can't work with the new cropduster field objects,
so we're going to have to ignore them in this context. This means that
reversion history will not contain cropduster images. We're hoping this
will not be a big deal.

Also there might be a nicer way to have reversion exclude these fields,
but I didn't readily see one, and didn't think it would be worth opening
an issue since we're on a fork that has not been kept up to date.